Format imports of Python/Cython files.

+++ python/subprojects/boosting/mlrl/boosting/cython/head_type.pxd
@@ -1,4 +1,4 @@
-from mlrl.common.cython._types cimport uint32, float32
+from mlrl.common.cython._types cimport float32, uint32
 
 
 cdef extern from "boosting/rule_evaluation/head_type_partial_fixed.hpp" namespace "boosting" nogil:
